# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vedika API
  version: 1.0.0
  description: |
    OpenAPI reference for the Vedika Express API.

    The implementation uses JSON request and response bodies, bearer-token authentication for protected routes,
    opaque cursor pagination for list endpoints, and full-replacement semantics for PUT operations.
servers:
  - url: http://localhost:3000
    description: Default local development server. The port is configurable at runtime.
...
paths:
...
  /booking-statuses:
    get:
      tags: [Booking Statuses]
      operationId: listBookingStatuses
      summary: List booking statuses
      description: Returns booking statuses in descending created-at order using cursor pagination.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paginated booking statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingStatusListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Booking Statuses]
      operationId: createBookingStatus
      summary: Create booking status
      description: Creates a new booking status.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '201':
          description: Booking status created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingStatusEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
...
  /event-statuses:
    get:
      tags: [Event Statuses]
      operationId: listEventStatuses
      summary: List event statuses
      description: Returns event statuses in descending created-at order using cursor pagination.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paginated event statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventStatusListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Event Statuses]
      operationId: createEventStatus
      summary: Create event status
      description: Creates a new event status.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '201':
          description: Event status created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventStatusEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
...
  /event-types:
    get:
      tags: [Event Types]
      operationId: listEventTypes
      summary: List event types
      description: Returns event types in descending created-at order using cursor pagination.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paginated event types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypeListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Event Types]
      operationId: createEventType
      summary: Create event type
      description: |
        Creates a new event type.

        The server also creates a default booking configuration for the event type, but the response only returns the
        event type itself.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '201':
          description: Event type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypeEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
...
  /service-providers:
    get:
      tags: [Service Providers]
      operationId: listServiceProviders
      summary: List service providers
      description: Returns service providers in descending created-at order using cursor pagination.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paginated service providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProviderListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Service Providers]
      operationId: createServiceProvider
      summary: Create service provider
      description: Creates a service provider. Email values are normalized to lowercase and blank optional strings become null.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceProviderRequest'
      responses:
        '201':
          description: Service provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProviderEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
...
  /call-records:
    get:
      tags: [Call Records]
      operationId: listCallRecords
      summary: List call records
      description: Returns call records in descending created-at order using cursor pagination. When `phoneNumber` is supplied, it is normalized to a canonical Indian 10-digit form and matched against both `callerNumber` and `receiverNumber`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: phoneNumber
          in: query
          required: false
          schema:
            type: string
          description: Indian phone number filter matched against caller or receiver after normalization.
      responses:
        '200':
          description: Paginated call records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallRecordListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Call Records]
      operationId: createCallRecord
      summary: Create call record
      description: Creates a call record. Caller and receiver numbers are normalized to a canonical Indian 10-digit form. `fileId` is optional and must reference an existing file when provided.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallRecordRequest'
      responses:
        '201':
          description: Call record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallRecordEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /files:
    post:
      tags: [Files]
      operationId: createFile
      summary: Create file
      description: Creates a file linked directly to an event booking. This endpoint requires `eventBookingId`; existing file usage in call records, payments, and voice notes is unchanged.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileRequest'
      responses:
        '201':
          description: File created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
...
  /calendar-events:
    get:
      tags: [Calendar Events]
      operationId: listCalendarEvents
      summary: List calendar events
      description: |
        Returns a merged, paginated event feed ordered by event date ascending.

        The feed includes followups whose `dueDate` falls inside the requested date range and event bookings whose
        `muhurat` falls inside the requested date range. Both `fromDate` and `toDate` are required and inclusive
        across the full UTC day.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: fromDate
          in: query
          required: true
          description: Lower inclusive date bound in `YYYY-MM-DD` format.
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          description: Upper inclusive date bound in `YYYY-MM-DD` format.
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated calendar events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /payments:
    get:
      tags: [Payments]
      operationId: listPayments
      summary: List payments
      description: Returns payments in descending created-at order using cursor pagination. When `serviceId` is supplied, it filters to payments for that exact service.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: serviceId
          in: query
          required: false
          schema:
            type: string
          description: Exact service-id filter.
      responses:
        '200':
          description: Paginated payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Payments]
      operationId: createPayment
      summary: Create payment
      description: Creates a payment linked to a service. `paymentProofFileId` is optional and must reference an existing file when provided.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /default-booking-configurations:
    get:
      tags: [Default Booking Configurations]
      operationId: listDefaultBookingConfigurations
      summary: List default booking configurations
      description: |
        Returns default booking configurations in descending created-at order using cursor pagination.

        `defaultStartTime` is always returned as a time-of-day string in `HH:mm:ss` format.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paginated default booking configurations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultBookingConfigurationListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
...
  /event-bookings:
    get:
      tags: [Event Bookings]
      operationId: listEventBookings
      summary: List event bookings
      description: |
        Returns event bookings in descending created-at order using cursor pagination.

        The `name`, `phoneNumber`, `fromDate`, and `toDate` filters can be combined. Date filtering uses overlap
        semantics, so any booking touching the requested range is included.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: name
          in: query
          description: Case-insensitive customer-name substring filter.
          schema:
            type: string
        - name: phoneNumber
          in: query
          description: Exact phone-number match against any stored phone-number field.
          schema:
            type: string
        - name: fromDate
          in: query
          description: Lower date bound in `YYYY-MM-DD` format.
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: Upper date bound in `YYYY-MM-DD` format.
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated event bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventBookingListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Event Bookings]
      operationId: createEventBooking
      summary: Create event booking
      description: |
        Creates an event booking.

        `serviceProviderIds` can be supplied on create. Each selected provider is linked to the booking and
        automatically synced to a backing `Service` row. The response does not currently return provider ids.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBookingRequest'
      responses:
        '201':
          description: Event booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventBookingEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /customer-interactions:
    get:
      tags: [Customer Interactions]
      operationId: listCustomerInteractions
      summary: List customer interactions
      description: |
        Returns customer interactions in descending created-at order using cursor pagination.

        The list can be filtered by linked event booking, ignored state, or only-unlinked records.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: eventBookingId
          in: query
          description: Only return interactions linked to the given event booking id.
          schema:
            type: string
            minLength: 1
        - name: ignored
          in: query
          description: Filter by ignored state.
          schema:
            type: boolean
        - name: unlinkedOnly
          in: query
          description: Only return interactions that are not linked to any event booking.
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated customer interactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerInteractionListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
    post:
      tags: [Customer Interactions]
      operationId: createCustomerInteraction
      summary: Create customer interaction
      description: |
        Creates a customer interaction.

        Omitted `eventBookingIds` becomes an empty array, and duplicate ids are deduplicated.
        `voiceNote` is optional and allowed only when `interactionType` is `WALK_IN`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInteractionRequest'
      responses:
        '201':
          description: Customer interaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerInteractionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /followups:
    get:
      tags: [Followups]
      operationId: listFollowups
      summary: List followups
      description: Returns followups in descending created-at order using cursor pagination and exact-match filters.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: dueDate
          in: query
          description: Exact due-date match as an ISO-8601 datetime string.
          schema:
            type: string
            format: date-time
        - name: type
          in: query
          description: Followup type filter.
          schema:
            $ref: '#/components/schemas/FollowupType'
        - name: eventBookingId
          in: query
          description: Exact event-booking-id match.
          schema:
            type: string
            minLength: 1
        - name: serviceProviderId
          in: query
          description: Exact service-provider-id match.
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Paginated followups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowupListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Followups]
      operationId: createFollowup
      summary: Create followup
      description: |
        Creates a followup.

        `BOOKING` followups require `eventBookingId`; `SERVICE` followups require `serviceProviderId`.
        `customerInteractionId` is optional for both variants.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowupRequest'
      responses:
        '201':
          description: Followup created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowupEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    LimitParam:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor from the previous page.
      schema:
        type: string
  responses:
...
    BadRequestResponse:
      description: Invalid request payload or query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedResponse:
      description: Missing, malformed, expired, or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundResponse:
      description: Requested resource or related resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictResponse:
      description: Request conflicts with an existing unique record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: Human-readable message
...
    PageInfo:
      type: object
      required: [limit, hasNextPage, nextCursor]
      properties:
        limit:
          type: integer
          minimum: 1
        hasNextPage:
          type: boolean
        nextCursor:
          type: string
          nullable: true
...
    FollowupType:
      type: string
      enum: [BOOKING, SERVICE]
...
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    BookingStatus:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingStatusEnvelope:
      type: object
      required: [bookingStatus]
      properties:
        bookingStatus:
          $ref: '#/components/schemas/BookingStatus'
    BookingStatusListEnvelope:
      type: object
      required: [bookingStatuses, pageInfo]
      properties:
        bookingStatuses:
          type: array
          items:
            $ref: '#/components/schemas/BookingStatus'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    EventStatusEnvelope:
      type: object
      required: [eventStatus]
      properties:
        eventStatus:
          $ref: '#/components/schemas/EventStatus'
    EventStatusListEnvelope:
      type: object
      required: [eventStatuses, pageInfo]
      properties:
        eventStatuses:
          type: array
          items:
            $ref: '#/components/schemas/EventStatus'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    EventTypeEnvelope:
      type: object
      required: [eventType]
      properties:
        eventType:
          $ref: '#/components/schemas/EventType'
    EventTypeListEnvelope:
      type: object
      required: [eventTypes, pageInfo]
      properties:
        eventTypes:
          type: array
          items:
            $ref: '#/components/schemas/EventType'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    ServiceProviderRequest:
      type: object
      required: [name, commissionRate]
      properties:
        name:
          type: string
          minLength: 1
        phoneNumber:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        commissionRate:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage commission rate from 0 to 100 inclusive. Decimal values are allowed.
...
    ServiceProviderEnvelope:
      type: object
      required: [serviceProvider]
      properties:
        serviceProvider:
          $ref: '#/components/schemas/ServiceProvider'
    ServiceProviderListEnvelope:
      type: object
      required: [serviceProviders, pageInfo]
      properties:
        serviceProviders:
          type: array
          items:
            $ref: '#/components/schemas/ServiceProvider'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    FileRequest:
      type: object
      additionalProperties: false
      required: [gcsPath, extension, eventBookingId]
      properties:
        gcsPath:
          type: string
        extension:
          type: string
        originalName:
          type: string
          nullable: true
        eventBookingId:
          type: string
...
    FileEnvelope:
      type: object
      required: [file]
      properties:
        file:
          $ref: '#/components/schemas/File'
    CallRecordRequest:
      type: object
      additionalProperties: false
      required: [callerNumber, receiverNumber]
      properties:
        callerNumber:
          type: string
          description: Phone number normalized to a canonical Indian 10-digit form before storage.
        receiverNumber:
          type: string
          description: Phone number normalized to a canonical Indian 10-digit form before storage.
        fileId:
          type: string
          nullable: true
...
    CallRecordEnvelope:
      type: object
      required: [callRecord]
      properties:
        callRecord:
          $ref: '#/components/schemas/CallRecord'
    CallRecordListEnvelope:
      type: object
      required: [callRecords, pageInfo]
      properties:
        callRecords:
          type: array
          items:
            $ref: '#/components/schemas/CallRecord'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    CalendarEvent:
      type: object
      required: [date, type, objectId]
      properties:
        date:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/CalendarEventType'
        objectId:
          type: string
    CalendarEventListEnvelope:
      type: object
      required: [events, pageInfo]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEvent'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    PaymentRequest:
      type: object
      additionalProperties: false
      required: [mode, amount, date, serviceId]
      properties:
        mode:
          $ref: '#/components/schemas/PaymentMode'
        amount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
          description: Positive decimal string with up to 2 fractional digits.
        date:
          type: string
          format: date
          description: Payment date in `YYYY-MM-DD` format.
        serviceId:
          type: string
        paymentProofFileId:
          type: string
          nullable: true
...
    PaymentEnvelope:
      type: object
      required: [payment]
      properties:
        payment:
          $ref: '#/components/schemas/Payment'
    PaymentListEnvelope:
      type: object
      required: [payments, pageInfo]
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    DefaultBookingConfigurationListEnvelope:
      type: object
      required: [defaultBookingConfigurations, pageInfo]
      properties:
        defaultBookingConfigurations:
          type: array
          items:
            $ref: '#/components/schemas/DefaultBookingConfiguration'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    EventBookingRequest:
      type: object
      required:
        [
          mode,
          bookingStatusId,
          eventStatusId,
          eventTypeId,
          bookingStart,
          bookingEnd,
          customerName,
          phoneNumber1
        ]
      properties:
        mode:
          $ref: '#/components/schemas/EventBookingMode'
        bookingStatusId:
          type: string
          minLength: 1
        eventStatusId:
          type: string
          minLength: 1
        eventTypeId:
          type: string
          minLength: 1
        bookingStart:
          type: string
          format: date-time
        bookingEnd:
          type: string
          format: date-time
        muhurat:
          type: string
          format: date-time
          nullable: true
        customerName:
          type: string
          minLength: 1
        phoneNumber1:
          type: string
          minLength: 1
        phoneNumber2:
          type: string
          nullable: true
        phoneNumber3:
          type: string
          nullable: true
        referredBy:
          type: string
          nullable: true
        serviceProviderIds:
          type: array
          description: Optional array of service-provider ids on create. Omitted or empty means no linked providers.
          items:
            type: string
            minLength: 1
...
    EventBookingEnvelope:
      type: object
      required: [eventBooking]
      properties:
        eventBooking:
          $ref: '#/components/schemas/EventBooking'
...
    EventBookingListEnvelope:
      type: object
      required: [eventBookings, pageInfo]
      properties:
        eventBookings:
          type: array
          items:
            $ref: '#/components/schemas/EventBooking'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    CustomerInteractionRequest:
      type: object
      required: [interactionType, occurredAt]
      properties:
        interactionType:
          $ref: '#/components/schemas/CustomerInteractionType'
        occurredAt:
          type: string
          format: date-time
        eventBookingIds:
          type: array
          description: Optional array of event-booking ids. Omitted becomes an empty array.
          items:
            type: string
            minLength: 1
        voiceNote:
          $ref: '#/components/schemas/CustomerInteractionVoiceNoteRequest'
...
    CustomerInteractionEnvelope:
      type: object
      required: [customerInteraction]
      properties:
        customerInteraction:
          $ref: '#/components/schemas/CustomerInteraction'
    CustomerInteractionListEnvelope:
      type: object
      required: [customerInteractions, pageInfo]
      properties:
        customerInteractions:
          type: array
          items:
            $ref: '#/components/schemas/CustomerInteraction'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    FollowupRequest:
      oneOf:
        - $ref: '#/components/schemas/BookingFollowupRequest'
        - $ref: '#/components/schemas/ServiceFollowupRequest'
...
    FollowupEnvelope:
      type: object
      required: [followup]
      properties:
        followup:
          $ref: '#/components/schemas/Followup'
    FollowupListEnvelope:
      type: object
      required: [followups, pageInfo]
      properties:
        followups:
          type: array
          items:
            $ref: '#/components/schemas/Followup'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'